requirements.txt
discord.py
python-dotenv
colorama
aioconsole